Add save/unsave feature

+++ Frontend/Plated2/core-model/src/main/java/com/priyanshparekh/core/model/recipe/ViewRecipeResponse.kt
@@ -1,7 +1,7 @@
 package com.priyanshparekh.core.model.recipe
 
 data class ViewRecipeResponse(
-    val recipeId: Int,
+    val recipeId: Long,
     val name: String,
     val cuisine: String,
     val category: String,
